add scale modal
